exteranlize: rolled back report data move
